requirements.txt
# Core dependencies for Supply Chain Analytics Platform
# Python 3.8+ required

# Data generation and extraction
Faker==22.6.0
pandas==2.1.4
numpy==1.24.3

# Snowflake connection
snowflake-connector-python==3.7.0
snowflake-sqlalchemy==1.5.1

# dbt (data transformation)
dbt-core==1.7.8
dbt-snowflake==1.7.2

# Machine Learning
scikit-learn==1.4.0
joblib==1.3.2

# Configuration
PyYAML==6.0.1

# Utilities
python-dotenv==1.0.1
tqdm==4.66.1

# Optional: Jupyter for notebooks
jupyter==1.0.0
ipykernel==6.29.0
matplotlib==3.8.2
seaborn==0.13.1
